Empty

+++ srcs/execution/excute.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   excute.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hnakayam <hnakayam@student.42.fr>          +#+  +:+       +#+        */
+/*   By: retanaka <retanaka@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/13 14:17:03 by hnakayam          #+#    #+#             */
-/*   Updated: 2024/11/21 15:59:44 by hnakayam         ###   ########.fr       */
+/*   Updated: 2024/11/22 15:37:41 by retanaka         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -258,6 +258,26 @@ int	create_word(char **pp, t_tkn *tkns, t_env *env, char *st)
 	return (1);
 }
 
+int	is_env_variable(const char *head, const char *tail)
+{
+	size_t	len;
+	size_t	i;
+
+	len = tail - head;
+	if (*head != '$' || len < 2)
+		return (0);
+	if (head[1] != '_' && !ft_isalpha(head[1]))
+		return (0);
+	i = 1;
+	while (i < len)
+	{
+		if (!ft_is_charactor(head[i]))
+			return (0);
+		i++;
+	}
+	return (1);
+}
+
 int	store_cmd(t_cmd *cmd, t_tkn *tkns, t_env *env, char *st)
 {
 	size_t	t_i;
@@ -271,21 +291,39 @@ int	store_cmd(t_cmd *cmd, t_tkn *tkns, t_env *env, char *st)
 	{
 		if (type_is_red(tkns[t_i].type))
 		{
-			cmd->reds[r_i].type = tkns[t_i++].type;
 			cmd->reds[r_i].file_fd = -1;
 			cmd->reds[r_i].std_target_fd = -1;
-			if (!create_word(&(cmd->reds[r_i].target), &tkns[t_i], env, st))
-				return (cmd->reds[r_i].type = TAIL, delete_cmd_exe(cmd), 0);
-			r_i++;
+			cmd->reds[r_i].type = tkns[t_i++].type;
+			cmd->reds[r_i].is_ambiguous = 0;
+			if (is_env_variable(tkns[t_i].head, tkns[t_i].tail) && !get_value(tkns[t_i].head, env, st))
+			{
+				cmd->reds[r_i].target = malloc(sizeof(char) * (tkns[t_i].tail - tkns[t_i].head + 1));
+				if (!cmd->reds[r_i].target)
+					return (cmd->reds[r_i].type = TAIL, delete_cmd_exe(cmd), 0);
+				cmd->reds[r_i].target[tkns[t_i].tail - tkns[t_i].head] = '\0';
+				ft_memmove(cmd->reds[r_i].target, tkns[t_i].head, tkns[t_i].tail - tkns[t_i].head);
+				cmd->reds[r_i++].is_ambiguous = 1;
+			}
+			else
+			{
+				if (!create_word(&(cmd->reds[r_i].target), &tkns[t_i], env, st))
+					return (cmd->reds[r_i].type = TAIL, delete_cmd_exe(cmd), 0);
+				r_i++;
+			}
 		}
 		else
 		{
-			if (!create_word(&(cmd->words[w_i]), &tkns[t_i], env, st))
-				return (delete_cmd_exe(cmd), 0);
-			w_i++;
+			if (!is_env_variable(tkns[t_i].head, tkns[t_i].tail) || get_value(tkns[t_i].head, env, st))
+			{
+				if (!create_word(&(cmd->words[w_i]), &tkns[t_i], env, st))
+					return (delete_cmd_exe(cmd), 0);
+				w_i++;
+			}
 		}
 		t_i += count_tkns_for_word(&tkns[t_i]);
 	}
+	cmd->words[w_i] = NULL;
+	cmd->reds[r_i].type = TAIL;
 	return (1);
 }
 
@@ -302,11 +340,9 @@ t_cmd	*expand_cmd(t_cmd *cmd, t_tkn *tkns, t_env *env, int status)
 	cmd->words = malloc(sizeof(char *) * (words_len + 1));
 	if (!cmd->words)
 		return (free(st), NULL);
-	cmd->words[words_len] = NULL;
 	cmd->reds = malloc(sizeof(t_red) * (reds_len + 1));
 	if (!cmd->reds)
 		return (free(st), free(cmd->words), NULL);
-	cmd->reds[reds_len].type = TAIL;
 	if (!store_cmd(cmd, tkns, env, st))
 		return (free(st), NULL);
 	return (free(st), cmd);
